Recategorize nuclide-react-native-common as a *-base
Summary:
`nuclide-react-native-common` requires `apm` packages (and uses Atom APIs directly - gasp!) but it's `testRunner` is `npm`. This diff renames it to `*-base` - the suffix for Node `apm` packages - and changes its `testRunner` to `apm`.

An upcoming lint rule is going to enforce that `npm` packages to do not require `apm` packages.

Reviewed By: matthewwithanm

Differential Revision: D3632888

fbshipit-source-id: 4f09b772046afeefafbd96ac96ee5d85adc06ca4