Fix non restoring remote files on Windows
Summary:
On MacOS, when Atom tries to open remote files from a serialized state it converts its path to something like: `nuclide:/host/path` (notice the single `/` after the `:`)
We have a logic that fixes that.

On Windows, however, it is converted to the format of `nuclide:\host\path` instead.
This diff adds the fix for this case as well.

Fixes #631, #619

Reviewed By: matthewwithanm

Differential Revision: D3634237

fbshipit-source-id: 5940beb918d729bea4fc3afca062df3eb1b79611