Decrease default scrollback
Summary:
The initial render of all those nodes can be very slow.

In the long term, we need to add windowing or load-on-demand behavior, but in the short term, this should alleviate the situation somewhat.

Note that this will help initial render, but probably not affect updates much. (That was addressed by D3585285.)

Reviewed By: yinghuitan

Differential Revision: D3633939

fbshipit-source-id: 02a228e4da904f5d301cdd7257e3242ad90fec2b