Fix UnsubscriptionError log spew in integration tests
Summary:
Every remote integration test with Mercurial fixtures currently has a bunch of unsubscription errors when the server is stopped. This is because we aren't properly cleaning up the corresponding `HgRepositoryClient`, leaving many dangling subscriptions that error out.

We can fix this by removing the project before stopping the server.

Reviewed By: johnislarry

Differential Revision: D3634181

fbshipit-source-id: 83175d7f4a7751e66a1406ca0c09c55cd8ac24c8