Add pending/resolved breakpoint state in chrome (7/N)
Summary:
This diff is required to support breakpoint_resolved notification from hhvm.
When user set a breakpoint in a file not loaded yet, we will get unresolved breakpoint response from hhvm which is forwarded to chrome. However chrome does not have the concept of pending breakpoint so whenever it got a response it *thinks* the breakpoint has been resolved and call _addResolvedLocation(). Then later when hhvm sends breakpoint_resolved notification to chrome, it believes hhvm binds to a new location, the result is that you will get two breakpoints in UI: one in original location, one in new bound location.
To fix this issue, I added an optional "resolved" field to "setBreakpointByUrl" API telling chrome that whether the breakpoint is resolved or not and skip _addResolvedLocation() if not resolved. For other backend not leveraging "resolved" field we still do the old behavior(call _addResolvedLocation()) so it will not regress other debug engine.

Note: later I may do more refactoring to chrome code but this is the simplest logic I can get this working.

Reviewed By: johnislarry, jgebhardt

Differential Revision: D3597669

fbshipit-source-id: 58c6569f2b0006a065996816ecec855f8efd243b