Report npm-to-apm requires as part of no-cross-atom-imports
Summary:
A package uses the `npm` `testRunner` is assumed to only run in Node (i.e. it only runs on the server). So they shouldn't require modules that use Atom/DOM APIs. This changes makes `no-cross-atom-imports` enforce that.

The packages that were violating the rule were fixed in D3634027, D3632888, D3632812, and D3633002.

Reviewed By: hansonw

Differential Revision: D3634794

fbshipit-source-id: 90dde0032db87d041903946fc00568edb98512d1