Create utility for converting React element into something that can be used with view registry
Summary:
Atom's ViewRegistry doesn't understand React elements. D3637244 provides a way to interop, but it requires a separate model for storing state. This utility allows you to keep your state in a React component which is often more convenient. It does this by mounting the React element and adding an `element` property to it which references the container node. (The `element` property is a special one to Atom's ViewRegistry). Because we're returning the mounted component, all instance methods on your component class are available, allowing you to implement Atom protocols like `getTitle()` on your component.

Like the tools in D3637244 and D3637269, this uses the ReactMountRootElement custom HTML element for the container.

See D3637425 for an example usage.

Reviewed By: ssorallen

Differential Revision: D3637409

fbshipit-source-id: fd88400382c903e2cccc913538507180194bbfb0