Fix failed evaluation case
Summary:
t12378872 sums the problem up nicely.  Here is my explanation for the fix:

The issue is that the errors we report in the backend from failed evaluation (for php) never make it to the
frontend.  On every evaluation request, the watch expression store tries to re-evaluate all its "active"
expressions.  This includes watch expressions that the user has typed into the watch window, as well as
console evaluation requests that have not received responses from the backend yet.  When the user types an
expression into the console that won't evaluate (e.g. due to a syntax error, for example `2 = 3`), the
frontend never receives the error response, and so then when the user does an action that causes watch
expressions to be re-evaluated such as stepping, or evaluating another expression in the console, the
unresolved expression with a syntax error is re-sent to the backend for evaluation, triggering yet another
error.

This diff solves this problem by handling this "failed evaluation" case correctly.  There are 3 steps:

1. In ConnectionMultiplexer, when the result object we get back from hhvm has an error, we make sure it has an error field set on the result object, because this is what the Chrome code expects.
2. In the chrome code, we make sure the callback that the NuclideBridge defines is passed the arguments it expects.
3. In the Bridge, when our promise rejects due to the failed evaluation, we convert _cachedSendCommand's null return value into a console-compliant, human-readable error message.

Reviewed By: yinghuitan

Differential Revision: D3646350

fbshipit-source-id: 8ad6e5168868c37f8338ef8d7d1819b76a4aa278