Add parser for Swift compiler commands
Summary:
When the Swift package manager builds a package, it stores the compiler commands
that are used to compile each of the files in the package. Those commands are
very interesting to us: we can provide them to SourceKit to get very precise
syntax highlighting, autocompletion, and more.

Add a parser for the llbuild YAML that contains the compiler commands, as well
as some tests.

Reviewed By: matthewwithanm

Differential Revision: D3636312

fbshipit-source-id: cf0c18382e42d7417e0bd42b3998ff6cef818e2e