Unsubscribe from debounced editor events in BreakpointDisplayController
Summary:
This is an Atom 1.9.0 fix that is 1.8.0 backwards-compat. In Atom 1.9.0, these events are emitted one more time after the editor has been disposed. Since they're never unsubscribed from, `gutter.decorateMarker` throws.

Also fixed the flow issue. The problem was that the `false` was meant as the first arg for `_update` (which expects an arg of type `number`). The flow error is cryptic.

Reviewed By: jgebhardt

Differential Revision: D3654397

fbshipit-source-id: 80429aa1255e2b55ffc7b446e772d9dc9beb418c