Do not cache or re-evaluate console expressions
Summary:
The WES will cancel and re-evaluate outstanding evaluation requests to the backend whenever re-evaluation is triggered (done by e.g. stepping, or evaling something in the console). This is fine for actual watch expressions, because if you need to re-evaluate, the watched expression could have changed.  It's not so good for console expressions, because we expect there to be a history of evaluated expressions in the console, so we don't want to cancel outstanding requests to the backend since we expect them to come back.

I was running into this when trying to trigger bp's from the console REPL.  There was an infinite loop where:

1. I would evaluate a function with a bp set in the body (this request won't come back until the user resumes the bp)
2. A request to eval would be sent to the backend, causing a change to RUNNING mode, and then back to PAUSE mode when the bp is hit
3. The debugger would request re-evaluation of watched expressions, as it always does when entering PAUSE mode
4. The WES would cancel the outstanding eval request, and issue a new one, going back to (2)

With this diff we no longer store the eval request as a watched expression, so it doesn't get re-evaluated at step (3) above.

This unblocks my work on triggering bp's from the console, but unfortunately this doesn't fix the analogous problem for watch expressions.  If you enter `foo()` as a watch expression, and `foo` has a bp set in its body, you get this same infinite loop problem.  One thing that might fix it is to extend the chrome protocol a little bit and mark the Debugger.pause message that gets sent back from the eval request as special, so that the WES knows not to trigger a re-evaluation of watch expressions.  I'm open to other suggestions.

Reviewed By: yinghuitan

Differential Revision: D3652729

fbshipit-source-id: 5b7930981d8cae9974882a89b240f9dcf3e1138f