Separate user breakpoint change from UI update change in BreakpointStore (11/N)
Summary:
There are several clients who are interested in the BreakpointStore changes: normal Nuclide UI is interested in any BreakpointStore changes to update UI, while Bridge.js is only interested in user action caused breakpoint changes to sync back to chrome and backend engine. We did not do the separation before which caused trouble, now I am fixing this.

Also, it used to sync the whole breakpoints list to chrome/backend when user changes single breakpoint. Now we are only sending one breakpoint across which should be an improvement for perf.

Reviewed By: johnislarry

Differential Revision: D3639211

fbshipit-source-id: 7dffa82512efecea83da1db6e072b00fcc11fc68