Use the consumed sync NUX service (if available) every time a NUX is dismissed.
Summary:
Use the consumed NUX service in the `NuxManager` whenever a NUX is completed.
The function from the activation class is bound to the activation class's execution context and passed into the manager, which is then invoked when a NUX is completed.
The function tries to use the consumed service if one exists.

Reviewed By: jgebhardt

Differential Revision: D3646079

fbshipit-source-id: 4765e8946b4717321e70f88b81e9d82f176d6e56