Stub electron typedefs
Summary: This diff creates a proper place for Electron's type definitions. So far it only covers the props off the electron module, `ipcRenderer` and `screen`. Since the exports of the module vary on who the caller is, the appropriate props are typed as optional. Also, the deprecated `ipc` module is removed (`ipcRenderer` replaced it).

Reviewed By: matthewwithanm

Differential Revision: D3660350

fbshipit-source-id: ecf0ef7663c0fa9fab155eb723dda3c2e064b531