Mark files without source in the PHP debugger
Summary:
This is pretty uncommon, but we can have source-less files in the PHP debugger.  The example I ran
into:

1. eval a function with a bp in the body
2. the callstack is something like: main -> include -> func

The `include` frame is there because hhvm wraps the eval'd code inside its own psuedo-main module.  There's no
source for the `include` frame.  So we're going to use the `hasSource` property to properly handle this in the
front-end (in the next diff).

Reviewed By: yinghuitan

Differential Revision: D3665243

fbshipit-source-id: 3e94018953e84415a5bd8f528d8048f87c92a3c4