Use tail position for datatip overlay instead of head plus left-adjust
Summary:
In D2875404, jgebhardt introduced the Nuclide datatip marker using
an Atom overlay marker with `position: 'head'` plus an adjustment
to the `left` style by the width of the selected range.

Directly messing with the `left` style causes some very strange issues
when the datatip hasn't yet been laid out. Depending on timing of the
reflow passes, it may or may not take effect, causing various issues:

1. Rapid flickering back and forth
2. Datatips laying out off the edge of the page
3. Strange behavior if the marker doesn't have a `head`

I believe the correct fix is just to use the `tail` of the marker,
rather than using the `head` plus adjusting to the left. From the docs:

https://atom.io/docs/api/v1.8.0/TextEditorMarker

> If you think of a marker as an editor selection, the tail is the part that's stationary and the head is the part that moves when the mouse is moved.

In this case, I believe we always want to attach the datatip to the `tail`,
since that's the part of the selection which remains stationary as the
mouse is dragged.

Reviewed By: jgebhardt

Differential Revision: D3678663

fbshipit-source-id: 9630b0abf8f6d696075cc49f64d37fa01ddbf90e