Clear the watched expressions on debugger stop
Summary:
The watch expression store will not clear its watched expressions map when the debugger is stopped.
The following scenario illustrates the problem:

1. User attaches debugger and watches an expression E
2. User stops the debugger and re-attaches it
3. The watch expression map isn't cleared, so we are still watching E, but this is not shown in the UI

At best, this will just cause an eval request to the backend for E every pause, and the user will be unaware
of this because the watch expression is not reflected in the UI.  At worst, E could have a side-effect, and be
defined in the scope that the user is debugging in (2), causing confusing and wrong behavior.

I noticed this extra eval request in the logs while debugging a separate problem, and clearing the watch
expression map on debugger stop should fix the issue.

Reviewed By: jgebhardt

Differential Revision: D3687681

fbshipit-source-id: e75176a60a6efe8dfc6d129cd6bb0535c5362f35