Fix flow integration test when tool-bar is at top
Summary:
This fixes `flow-diagnostics-gutter-integration-spec` and `flow-diagnostics-gutter-remote-integration-spec` when the tool-bar is at the top. When the tool-bar is taking up vertical space, the gutter diagnostic is too far below the fold, so Atom never adds it to the DOM.

I couldn't get the editor scroll down in the test. So I just made the fixture shorter, so that the gutter diagnostic is a bit further up.

Reviewed By: nmote

Differential Revision: D3695983

fbshipit-source-id: b4b9fd82262b4f732b64bc4627232687c7853939