Clear remote connection configs when server is shutdown
Summary:
If we shut down the server and reload Atom, we shouldn't even try reusing the config.
This happens on upgrade, when you remove a project and click Shutdown, or when you use "Kill Nuclide Server And Restart".

The main purpose of this is to allow logging of when a reconnect fails (Currently it'd yell for everyone on release.)

Reviewed By: zertosh

Differential Revision: D3693796

fbshipit-source-id: 421151114f288abea3ba07d2e05d4830b935c20e