Remove handleGrammarUsed triggering when building own text editors
Summary: The idea behind calling `atom.workspace.handleGrammarUsed` was to trigger the relevant activation hooks. This was a regression introduced in Atom 1.9.0, which was later fix in Atom 1.9.8 (https://github.com/atom/atom/pull/12375). Now, just adding the newly built text editor to the registry, is enough to trigger the activation hooks.

Reviewed By: hansonw

Differential Revision: D3696037

fbshipit-source-id: 6e40b36c6ca757fa53b3f90ee099409500b81e92