Fix transpiler sourcemap filenames
Summary: This fixes source map filenames showing up as "unknown". Since the source maps embed the filename, the cache key now has to account for that. That means that multiple copies of Nuclide can't share the same cache. However, because we mostly develop without source maps, there's code to handle that case, so you can still share the cache in those cases.

Reviewed By: yinghuitan

Differential Revision: D3720299

fbshipit-source-id: 36c82261d07faf098069896bcfcf2fbd4b664fe7