Fake script parse for breakpoint
Summary: This is a regression when I improve the breakpoint binding experience: we used to sync all breakpoints across IPC for each single breakpoint add/remove which is a huge perf issue. I changed to only send breakpoint across for the changed breakpoint. However I forgot to do the fake script parse hack which is added in this diff.

Reviewed By: johnislarry

Differential Revision: D3723353

fbshipit-source-id: 782269c9ee6183f2588a57944fd7e1b6a84695fe