Use ConnectableObservable in nuclide-rpc.
Summary:
Before this diff, there were subtle differences between local and remote
observables in remote interfaces. This diff makes local and remote observables
have closer, but not identical, semantics.

This change replaces all uses of Observable in rpc interfaces to ConnectableObservable.
ConnectableObservables subscriptions share the underlying subscription.
To convert a regular Observable to a ConnectableObservable call publish().
On the calling side, you must call connect() on a ConnectableObservable before
it will begin sending values. The easiest way to handle that is to use .refCount()
on the ConnectableObservable and then use it as you would normally.

Note that observable.publish.refCount() === observable.share().

Use of Observable in an rpc interface results in an error directing you to use
ConnectableObservable.

The outstanding differences are:

- if the original call to create the observable throws, then in the local case, the caller will see the throw, in the remote case an observable which throws is seen.
- in the local case, an observable returned over an rpc interface may be connected to, unconnected, then reconnected. In the remote case the attempt to reconnect will through.

TLDR

Use ConnectableObservable instead of Observable in RPC interfaces
Use .publish() to convert an Observable to ConnectableObservable
Use .refCount() when consuming a ConnectableObservable

Reviewed By: nmote

Differential Revision: D3671250

fbshipit-source-id: 7e263f98330263172a1db310c937172e799e2c04