Refactor nuclide-task
Summary: By using the `--require` flag to load the transpiler, we're guaranteed to have it loaded before we start the bootstrap script. Since the bootstrap script doesn't have to deal with starting the transpiler anymore, we can remove a lot of code from it and even transpile the bootstrap script itself!

Reviewed By: hansonw

Differential Revision: D3728339

fbshipit-source-id: 75f65f98ff89e6fd9208cd0be37c2146b14130a6