Force the NUX trigger type to be non null
Summary:
Set the trigger type to only be `editor` and accept more cases as they are developed (also removed my TODO :P)
This makes it easier to understand when reading the docs, as the `trigger` property is nullable anyway. If it is non-null, then the trigger type should have a value.

Reviewed By: jgebhardt

Differential Revision: D3731934

fbshipit-source-id: 5931b9cb319ed9dc8ab6ee88bd097f8bc5242e3f