Improve modal component dismissal behavior
Summary:
Previously, we were listening to the "blur" event, however actions caused by interacting with the modal can trigger elements outside of it to focus. So, for example, if you would start a logger process from the console source modal, a new panel would be added and its item focused, therefore dismissing the modal.

This new behavior listens for a click outside of the modal. It also listens for the "cancel" command anywhere in the document (instead of just on the modal element).

We need a zero timeout here to make sure that we don't accidentally capture the same click that showed the modal in the first place.

Reviewed By: jgebhardt

Differential Revision: D3736832

fbshipit-source-id: 241b1f8f309a5f53d6d309a8da3276fc136764c4