Create next and end buttons in NUXes to differentiate ending/moving to next NUX
Summary:
As discussed offline with jgebhardt , created buttons to differentiate between moving to the next NUX and skipping the whole tour.

The NUXes have custom styling instead of using UI buttons as the buttons have inconsistent styling between themes.
The next (Complete) button and end (Dismiss) buttons are always visible.

However, the next button is only "enabled" (via styles + click handler) iff there is no action to be completed to progress (e.g. button must be clicked) and there is a NUX to move to i.e. it we are not on the last NUX in the tour. Added a hint for this as well (see demo)

The "Dismiss" button turns into a "complete" button on the last NUX.

Also, the event listener is removed from the tooltip -- i.e clicking the tooltip no longer progresses to the next NUX.

Changelog: Added NUX navigation buttons

Reviewed By: jgebhardt

Differential Revision: D3728664

fbshipit-source-id: 359c62a8efa5d7d46643754a791bf1ce827bb262