Refactor NativeDebuggerService to remove DebuggerConnection
Summary:
Originally I introduced DebuggerConnection thinking that we might be able to support one IDE debugging multiple C++ connections using single RPC service.
But that introduces complexity in RPC and preventing code reusing.
This diff removes DebuggerConnection so that we can reuse the common code in future for all debuggers.

Reviewed By: johnislarry

Differential Revision: D3739238

fbshipit-source-id: 5dca85221b2bfee2f006c2bc3b4e55b8f0eef0ba