Add shellParse wrapper around shell-quote's parse
Summary:
`shell-quote`'s parse has a feature where it allows parsing of pipe operators.
These show up as objects with an `op` property in the result.

None of our callsites actually want this (in fact this will cause weird runtime errors further down the stack), so add a wrapper that throws if this happens.

Reviewed By: zertosh

Differential Revision: D3756537

fbshipit-source-id: e9218e27765d1ab62e56bdb8df5a798c29a04d86