Fix file-tree scrolling on lengthy loading
Summary:
Background:
File-tree supports opening deeply nested files, such that don't have their parents loaded yet. When performed, the file and all of its (yet unfetched) parents are added to the tree and the async fetch is scheduled. File-tree also supports automatically revealing files on switch.
The combination of the two above features required adding a notion of a `tracked` node. This kind of node is maintained in the view despite structural changes of the tree, so that when a deeply nested file is revealed it wouldn't be immediately hidden once the parents finish loading. Basically, so long as there is such a node, the file tree constantly scrolls to it.
The `tracked` node would be cleared when all loading is done.

The problem:
Unfortunately, sometimes the loading might fail and it takes a whole minute to timeout. Or, due to unexpected bugs, such as one that was fixed in D3739324, the loading of some folder might not finish at all. Then the `tracked` node is never cleared, which makes it impossible to scroll the file tree out of that node (it just jumps back).

This diff clears the tracked node when it detects a scroll that was performed non-programmatically. To detect such, it uses a flag that is set on a programmatic scroll request. This flag makes the next scroll event assume it was programmatic and not user-initiated. The flag is cleared each time scroll event is handled.

Reviewed By: matthewwithanm

Differential Revision: D3751800

fbshipit-source-id: fc03371b674eed1e6629b2e9bff67702a1fa86c0