Debugger Eval #1 -- Prevent caching of evaluations
Summary:
This diff extends the evaluation logic in the Bridge so that we use unique id's to identify evaluation requests/responses.  Because of this, the Bridge won't do any extra caching of evaluation requests, which is good for use cases like evaluating functions with bp's in their body -- where the result won't come back until at least one debugger pause event is emitted.

I know that adding this logic into the Bridge is not great -- so the next diff in the stack D3699243 refactors all of this logic out of the Bridge.

Reviewed By: yinghuitan

Differential Revision: D3688065

fbshipit-source-id: 0aa37358651b9e3f615288b0ed933fd9018efda0