Don't use promises of processes in observeProcess
Summary:
In working on D3759657, I realized that we have potential race conditions between a process's error event and the resolution of the promise containing that process. I think the best way forward is to just not accept promises of processes anymore. Since observables can be synchronous, those will work better. For example, instead of this:

  observeProcess(async () => { await doSomething(); return spawn(...); })

we can do this:

  Observable.fromPromise(doSomething()).switchMap(() => observeProcess(() => spawn(...)))

This mostly wasn't a big deal now that we're no longer using the createExecEnvironment util; the only thing actually relying on passing a Promise was the ClangServer.

Note that I didn't rewrite RpcProcess to be more idiomatic in its observable usage; instead I aimed for the minimal change. I do think it could be improved by refactoring the side-effects in `_ensureProcessImpl` into transforms though.

Reviewed By: hansonw

Differential Revision: D3763133

fbshipit-source-id: 9c8a17407657f31afce205336d016b662700699a