Guard against trying to observe exited processes
Summary: As I mentioned in D3763133, it's possible to create a process, for its error (or exit) event to be fired, and then to pass it to a utility like `observeProcess`. Disallowing promises means we make it far less likely, but it's still possible. I don't think there's any way around that, but we can do our best to detect the situation, and that's what this does (:

Reviewed By: hansonw

Differential Revision: D3765465

fbshipit-source-id: af9860c08cb0a33c2b90bb8c0e7c6f4ee81924f7