Remove command-alt-r shortcut
Summary:
The React Native packager uses this as a global reload shortcut. It's not a huge deal since it intercepts the calls, but it might be confusing for it to sometimes open the Recent Files Quick Open panel and sometimes not.

Changelog: Remove Recent Files keyboard shortcut due to conflict with React Native packager.

Reviewed By: nmote

Differential Revision: D3775140

fbshipit-source-id: efc7cc2ed836ccf6e0cf4837e5d3bbb795f0b75b