Buck: helper to generate diagnostics from Buck output
Summary:
Unfortunately Buck does not provide any sort of detailed compilation error data via the JSON websocket.
Until it does, we can parse the log output and look for patterns that match errors ourselves.
This is tailored to Clang/g++ output at the moment, but could be expanded to other backends as well.

Reviewed By: matthewwithanm

Differential Revision: D3768831

fbshipit-source-id: 62e30e51e2bcb05ad8209922a24f0697917d2b0d