Exclude identical suffixes in original/replacement from autofix
Summary:
This is a good idea in general because if the original and replacement share a
common suffix, the range for the autofix is overbroad. This means that edits in
the area that will not change when the autofix is applied will actually cause
the autofix to be invalidated.

Changelog: Make arc lint autofix more robust

Reviewed By: peterhal

Differential Revision: D3778557

fbshipit-source-id: 49a0a3f04ade51e383c52751ac774a39f686cf1c