Allow breakpoints to bind in files with no file extension
Summary:
The chrome model we use requires that we have "source parsed" for a file before we set a breakpoint
for it.  This is not good, and we have a hack to work around it where we spoof the message it expects.  We
keep a whitelist of things that are okay to spoof, but files with no extension weren't part of this
whitelist.

Reviewed By: jgebhardt

Differential Revision: D3789610

fbshipit-source-id: 011deeafc1d419ca2fb6fb72964f29f8d1b84b9e