Clang: Find local references implementation
Summary:
Clang provides an awesome `clang_findReferencesInFile` implementation for us.
You won't find it in their official docs, but it's in the code at:
https://github.com/llvm-mirror/clang/blob/840ee0d0622586ef77295a5f0a1000494a9a1f48/include/clang-c/Index.h#L5386

The one listed caveat is that it can't find non-argument usages in macros, e.g. if you do

```
int x = 0;
MY_MACRO;
```

then the reference to `x` would not be found. I don't think there's a good solution to this, anyway, so let's leave it for now.

Reviewed By: peterhal

Differential Revision: D3798866

fbshipit-source-id: a60f11f4d3fd052983aff162d3b13c1bf5037286