Fix lldb attach dialog selection bug
Summary:
There are two ways to set the selected process from the attach dialog.  One way is to explicitly
click the row in the table you want to select.  The second is to wait for a refresh of the process list
(this happens periodically, every second or so) and this will select a ProcessInfo object if you already had
one selected, using some logic.  We weren't using this logic in the first scenario, so we could sometimes have
bogus debugger targets.

Reviewed By: mostafaeweda

Differential Revision: D3824694

fbshipit-source-id: 89daf963a5fbd1266ae6431dc5d602c5b5f29c7b