Fix bug where booleans aren't showing up in locals window during PHP debugging
Summary:
In PHP debugging boolean values are showing up everywhere except the locals window.  This is because
we miss some normalization logic, usually done for the console, watch window, and datatips.  I'd like to
refactor this normalization logic to make the stringification happen at the last moment inside this react
component... but this diff is enough to unblock the release.

Facbook:

Changelog: Booleans display correctly in the locals window of the PHP debugger.

Reviewed By: mostafaeweda

Differential Revision: D3825021

fbshipit-source-id: 10f47876d0b70f052d4198c30a43078ed3f022b1