Make file tree collapse command global
Summary:
I used to be content to click on the file tree first to give it focus. But some
time ago, we implemented the temporary tabs thing, which meant that every time
you click on a file name it's opened in the editor and the file tree loses
focus. Which means to use this shortcut you need to find a directory and click
on that (which expands/collapsed it in the process).

This has frustrated me since and I'm finally fixing it.

Reviewed By: shushz

Differential Revision: D3824922

fbshipit-source-id: def22aa8327c2a1804ecf9b02b54e5871d2da2d9