Diff Section Navigation [4/n]: Use the diffSections state to render the DiffNavigationBar
Summary:
This diff uses the section computation logic refactored and set as state in the DiffViewComponent to render the navigation bar elements - also, in preparation to be consumed by another React component.
In this process, this reduces the number of DOM elements / React components created for the DiffNavigationBar to be the number of changed sections instead of being number of lines.
This makes the Diff View updates much faster and consumes less memory / elements, especially for files with many changes..

Reviewed By: jgebhardt

Differential Revision: D3815105

fbshipit-source-id: e49b1d86b1760c66ef73d5c79731c971ea274cd8