Use smartlog expression in revision timeline query
Summary:
Since the revision timeline now consumes a part of the smartlog data, while the in-prototype smartlog view need the smartlog data,
It makes sense to share the data by extracting the piece needed for the revision timeline and expose an API to access and listen to smartlog data changes.
This diff introduces the smartlog query and extracts the revision timeline data from it.
Following diffs will rename, restructure and introduce APIs to consume the smartlog data.

Reviewed By: jgebhardt

Differential Revision: D3812136

fbshipit-source-id: 4afce79edc4623c80e390d23094c41463d9fa367