Support eval breakpoints
Summary:
The idea is that the xdebug `eval` command will send us a response with the `status` attribute set to
`break`, and when we see this, we need to assume that we've hit a breakpoint in the code that was evaled, and
update the UI, by sending a resume followed by another break.

Things to consider for the future are putting the state machine into 'single request mode' when this happens,
so that it does not context switch to a different request, which would be a bad experience for the user.

Reviewed By: jgebhardt

Differential Revision: D3313259

fbshipit-source-id: 0fe2416ca367b07f13e1911f36b66844ef47e41a