Handle the dummy connection in the ConnectionMultiplexer
Summary:
This allows us to trigger bp's from the dummy connection.  It's real real dope.

The tricky part of this diff is management of whether the dummy connection is visible to the user.  It adds two new statuses to the state machine, used by the dbgp socket to communicate if the dummy connection should be shown in the UI.  This is determined by looking at if there are any pending eval requests that have hit breakpoints.

- If there are NOT, then the dummy connection is currently paused at the xdebug_break() call in the entry point, and so we should not allow it to become the enabled connection.
- If there ARE, then the dummy connection is paused outside of this entry point due to a previous eval request, and so we should allow it to become the enabled connection.

Reviewed By: jgebhardt

Differential Revision: D3764862

fbshipit-source-id: ee751e96c3674848891c48cde28832a8abab4b2f