Workaround possible deferred activation
Summary:
When doing an OSS upgrade, it is possible to get stuck in the deferred module state. This happens because upon upgrading, the package metadata is re-read (so new version number), but the main module is still the old code (so possibly no dummy deserializer). So the deferred cache key is written for the new version but against old code.

The fix is to simply reset the defer key. However, this still means that the first time you reload, you're thrown into the bad state. By moving `setUseLocalRpc`, the worse that can happen by being in the bad state is borked deserialization - instead of a red box with the message "Must call setUseLocalRpc before getService".

Fixes #698.

Reviewed By: nmote, mostafaeweda

Differential Revision: D3839909

fbshipit-source-id: ea9d2999f3397896a65b44c474aba28d46f699c0