Start of customizable prefedefined RPC types.
Summary:
This stack enables enables a custom set of predefined types in nuclide-rpc.

Currently the only type with a custom marshaller is NuclideUri, which is special
cased all through the nuclide-rpc package.

This stack removes any dependency on NuclideUri from nuclide-rpc by enabling a
custom set of predefined types to be specified by nuclide-rpc clients.

This diff adds a set of predefinedTypes to the ServiceParser constructor.
Moving the special casing of NuclideUri out of the ServiceParser.

Reviewed By: hansonw

Differential Revision: D3835147

fbshipit-source-id: d383496d60ea67ba55ae9f43918d3a31295a92af