Preserve original login environment and use it when launching Buck
Summary:
When the Nuclide server is first spawned, it may be with a custom environment/bashrc.
This irreversibly damages the environment, which we cannot replicate by spawning a new shell.
Buck is particularly sensitive to this, as it restarts if the environment changes. This causes slow builds if a user uses both the CLI and Nuclide.

Instead, preserve the original SSH environment, pass it into the Nuclide server, and use it when running Buck.

greenparty

Reviewed By: zertosh

Differential Revision: D3843040

fbshipit-source-id: f076830d7421ca009d6c17606f47c2612e13e528