Create a script that does `atom --wait --add` properly.
Summary:
This is a workaround for https://github.com/atom/atom/issues/1433 that provides
a script that is useful for the following items:

* `$HGEDITOR`
* `$GIT_EDITOR`
* `arc set-config editor`

To make this script work, I had to update `pkg/nuclide-remote-atom-rpc/bin/atom`
to return different exit codes for different types of failures.

The new script, which is designed for the use cases above,
is available at `pkg/nuclide-remote-atom-rpc/bin/atom-wait`. As explained in the
comments in the script, it takes a single argument and tries the following, in order:

1. Opens the file in a new tab in an existing Atom window. Does not return until the tab is closed.
2. Opens the file using `$ATOM_BACKUP_EDITOR`. This happens if Atom is not already open.
3. Opens the file using `$EDITOR`.

For people who have Atom/Nuclide open all the time and would like to use it for things
like drafting commit messages or interactive rebase, this is a nice option. The script
includes comments that explain how to set it up either locally or remotely.

Reviewed By: peterhal

Differential Revision: D3868204

fbshipit-source-id: a0381b4207201c823f826ca4c0dde0a4c55adaeb