Buck: Refactor BuckProject into a stateless BuckService
Summary:
`BuckProject` is an object that only contains a single `path` property, so it could easily just be a set of functions.

This is better for a few reasons:

- Anyone who wants to run a Buck command has to instantiate a local and remote object
- Remote objects need to be explicitly disposed, which makes code inconvenient
- Remote objects have an uncertain lifetime. Attempting to cache them is a big anti-pattern, as they become invalidated as soon as the server connection is released (which we have no control over).

This refactors `BuckProject` into a set of stateless global functions (all taking an extra root path parameter) and modifies the callsites accordingly.

A bunch of the unit tests had to be reworked - instead of mocking a `BuckProject` object, now we can just directly mock the `BuckService` export.

This results in a net -100 LOC since we no longer have to manage these `BuckProject` objects.

Reviewed By: matthewwithanm

Differential Revision: D3867227

fbshipit-source-id: a44a0fd82f6556229616f35ee3b6444e70dd991d